requirements.txt

ollama

faiss-cpu

langchain

langchain-community

python-dotenv

pandas

pypdf

streamlit

langchain-ollama

langchain-text-splitters

bs4

tqdm

pymupdf

wikipedia

langchain-openai

chromadb

sentence_transformers

langchain_huggingface

langchain_chroma

duckdb

langchain-groq

duckduckgo_search==5.3.1b1

mysql-connector-python

SQLAlchemy

validators==0.28.1

youtube_transcript_api

unstructured

pytube

numexpr

huggingface_hub

